requirements.txt
Flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
psutil>=5.9.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
playwright>=1.40.0
chromadb>=0.5.5
sentence-transformers>=2.7.0
tiktoken>=0.7.0
pywinpty>=2.0.0 ; sys_platform == "win32"